requirements.txt
openai>=1.0.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
PyMuPDF>=1.23.0
numpy>=1.21.0 